Update README.md
Updated latest date to drop with a 'W' as found in the 2021 academic calendar.  I also added an explanatory line to the description of the miscellaneous assignment clarifying (based on what the TA told me) that this assignment counts as a regular homework, not as an alternative, subject to the same, drop policy.